+++ Extension/src/background/api/filtering-log.ts
@@ -75,6 +75,7 @@ export type FilteringLogEvent = {
     cookieName?: string,
     cookieValue?: string,
     isModifyingCookieRule?: boolean,
+    cspReportBlocked?: boolean,
     replaceRules?: FilteringEventRuleData[],
     stealthActions?: StealthActionEvent['data']['stealthActions'],
 };