+++ Extension/src/background/services/filtering-log.ts
@@ -33,6 +33,7 @@ import {
     JsInjectEvent,
     ReplaceRuleApplyEvent,
     StealthActionEvent,
+    CspReportBlockedEvent,
 } from '@adguard/tswebextension';
 
 import { messageHandler } from '../message-handler';
@@ -123,6 +124,11 @@ export class FilteringLogService {
         defaultFilteringLog.addEventListener(FilteringEventType.JsInject, FilteringLogService.onScriptInjection);
 
         defaultFilteringLog.addEventListener(FilteringEventType.StealthAction, FilteringLogService.onStealthAction);
+
+        defaultFilteringLog.addEventListener(
+            FilteringEventType.CspReportBlocked,
+            FilteringLogService.onCspReportBlocked,
+        );
 
         if (UserAgent.isFirefox) {
             defaultFilteringLog.addEventListener(
@@ -348,12 +354,25 @@ export class FilteringLogService {
      * @param event Event with type {@link ReplaceRuleApplyEvent}.
      * @param event.data Destructed data from {@link ReplaceRuleApplyEvent}:
      * tab id, event id and stealthActions - last one is the bit-mask
-     * of applied {@link StealthActions} from webextension.
+     * of applied {@link StealthActions} from tswebextension.
      */
     private static onStealthAction({ data }: StealthActionEvent): void {
         const { tabId, eventId, stealthActions } = data;
 
         filteringLogApi.updateEventData(tabId, eventId, { stealthActions });
+    }
+
+    /**
+     * Records the blocked csp report.
+     *
+     * @param event Event with type {@link CspReportBlocked}.
+     * @param event.data Destructed data from {@link CspReportBlocked}:
+     * tab id, event id and cspReportBlocked - last one is a boolean flag.
+     */
+    private static onCspReportBlocked({ data }: CspReportBlockedEvent): void {
+        const { tabId, ...eventData } = data;
+
+        filteringLogApi.addEventData(tabId, { ...eventData });
     }
 
     /**